chapter-11/03-data-reprocessing/02-batch-reprocessor-stateful-apache-spark-delta-lake/requirements.txt
pyspark==4.0.0
delta-spark==4.0.0
